Parser was added and new implementations

+++ spec/src/main/java/org/eustrosoft/handlers/cms/dto/ResponseBlock.java
@@ -11,7 +11,7 @@ public interface ResponseBlock<T> {
 
     String getR();
 
-    Short getE();
+    Long getE();
 
     String getM();
